Fixed some issues with date and integer output for Xml.
git-svn-id: svn+ssh://developer.mitre.org/svnroot/giscore/trunk@25 5c99b798-f2dd-11dd-9a12-8f84353f4b73

+++ giscore/src/org/mitre/giscore/output/esri/GdbOutputStream.java
@@ -629,9 +629,9 @@ public class GdbOutputStream extends StreamVisitorBase implements
 
 	private Variant makeIntegerVariant(String name, int ft, Object value) {
 		if (value instanceof Number) {
-			return new Variant(name, ft, ((Number) value).longValue());
+			return new Variant(name, ft, ((Number) value).intValue());
 		} else if (value instanceof String) {
-			return new Variant(name, ft, new Long((String) value).longValue());
+			return new Variant(name, ft, new Integer((String) value).intValue());
 		} else {
 			throw new IllegalArgumentException("Uncoercible type found "
 					+ value.getClass());
